pyproject.toml
[project]
name = "mt5-trading-system"
version = "0.1.0"
description = "Broky + Metty: Kappa-powered algorithmic trading system for XAUUSD"
requires-python = ">=3.11"

dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pandas-ta>=0.3.14b1",
    "aiosqlite>=0.20.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "python-telegram-bot>=21.0",
    "mt5linux>=0.1.8",
    "matplotlib>=3.8.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
ml = [
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
broky = "broky.cli:main"
metty = "metty.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"